+++ changelog.md
@@ -6,7 +6,8 @@
 
 ### Unreleased
 
-There are currently no unreleased changes
+ - Fix slow scrolling on Linux and Windows
+ - On Windows and Linux, scroll horizontally when shift is pressed
 
 ### Release 0.5.3, 2017-05-02